Revamp bashrc and move aliases to .bash_aliases
A number of changes here. The primary one is moving aliases to a
dedicated file. Some aliases for programs unlikely to be installed on a
bunch of systems only activate if such a program can be found.

Also changes the PSI shell prompt to use a generator function instead of
an increasingly complicated string. This new one shows the current git
branch and a :) or :/ emoticon based on whether the last command
succeeded.